+++ cmd/gateway.go
@@ -28,7 +28,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/juicedata/godaemon"
 	"github.com/juicedata/juicefs/pkg/chunk"
 	"github.com/juicedata/juicefs/pkg/fs"
 	"github.com/juicedata/juicefs/pkg/meta"
@@ -185,33 +184,6 @@ func (g *GateWay) NewGatewayLayer(creds auth.Credentials) (minio.ObjectLayer, er
 			Umask:       uint16(umask),
 		},
 	)
-}
-
-func makeDaemonForSvc(c *cli.Context, m meta.Meta) error {
-	var attrs godaemon.DaemonAttr
-	logfile := c.String("log")
-	attrs.OnExit = func(stage int) error {
-		return nil
-	}
-
-	// the current dir will be changed to root in daemon,
-	// so the mount point has to be an absolute path.
-	if godaemon.Stage() == 0 {
-		var err error
-		attrs.Stdout, err = os.OpenFile(logfile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-		logger.Infof("open log file %s: %s", logfile, err)
-		if err != nil {
-			logger.Errorf("open log file %s: %s", logfile, err)
-		}
-	}
-	if godaemon.Stage() <= 1 {
-		err := m.Shutdown()
-		if err != nil {
-			logger.Errorf("shutdown: %s", err)
-		}
-	}
-	_, _, err := godaemon.MakeDaemon(&attrs)
-	return err
 }
 
 func initForSvc(c *cli.Context, mp string, metaUrl string) (*vfs.Config, *fs.FileSystem) {
